Update to work with the changed Pushy API.
It is no longer needed (or possible) to explicitly `connect()` an
`ApnsClient`; the connection state is managed by Pushy itself. This
means that is necessary to configure the APNS server and credentials
in the `ApnsClientBuilder` before `build()` can be called.

This means there is no longer any point in having a no-argument
arity of `make-client`, and the other arities need to take a `host`
argument. The `connect` function is also obsolete, and has been
removed.